Creating an individual class for the MyInfo Component
