+++ src/components/MyInfo.js
@@ -1,0 +1,20 @@
+import React from "react";
+
+function MyInfo() {
+  return (
+    <div>
+      <h1 id="header1">Joseph Schlitz</h1>
+      <p>
+        I am a software engineer at a small telecommunications company. Soon I
+        will be a software engineer 2 at the Space Telescope Science Institute.
+      </p>
+      <ul>
+        <li>Japan</li>
+        <li>Germany</li>
+        <li>Spain</li>
+      </ul>
+    </div>
+  );
+}
+
+export default MyInfo;
